+++ ntex-http/CHANGES.md
@@ -1,9 +1,13 @@
 # Changes
 
-## [0.1.0] - 2022-06-20
+## [0.1.0] - 2022-06-26
+
+* impl PartialEq for HeaderMap
 
 * impl TryFrom<&str> for Value
 
 * impl FromIterator for HeaderMap
+
+## [0.1.0] - 2022-06-20
 
 * Move http types to separate crate